requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.3
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
idna==3.10
imageio==2.35.1
imgaug==0.4.0
importlib-resources==6.4.5
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lazy-loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
packaging==25.0
pandas==2.0.3
pillow==10.4.0
polars==1.8.2
psutil==7.1.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.3
requests==2.32.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2023.7.10
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
triton==3.0.0
typing-extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.205
ultralytics-thop==2.0.17
urllib3==2.2.3
zipp==3.20.2
